rag/pyproject.toml
[project]
name = "tessera-rag"
version = "0.1.0"
description = "RAG system for mapping natural language to EQ DSP parameters"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "openai>=1.30.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "datasets>=2.20.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "pandas>=2.2.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tessera_rag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
